Flutter Login Screen Project
This Flutter project is a simple login screen that demonstrates the basics of user authentication. The application uses a hardcoded list of users and validates their credentials (email and password) through a simple login form.

Features
User Authentication: Allows users to log in using an email and password.
Secure Password Handling: Includes a toggle to show/hide the password input for security.
Responsive UI: The UI is designed to be responsive and works on different screen sizes.
Error Handling: Provides feedback if the user enters incorrect credentials.
Project Structure
main.dart: The entry point of the application. It initializes the app and sets up the main theme and home page.
HomePage: The main screen of the app where the login form is located. It includes text fields for email and password, and a login button.
Users Class: A simple class that defines the structure of a user object, containing an email and password.